% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching-archive.R
\name{fetch_updating_resource}
\alias{fetch_updating_resource}
\title{Fetch or read cache of a resource that potentially updates over time}
\usage{
fetch_updating_resource(
  fetch,
  check_response,
  cache_file_prefix = NULL,
  cache_invalidation_period = as.difftime(1L, units = "days"),
  force_cache_invalidation = FALSE,
  silent = FALSE,
  make_ancestor_directories = TRUE
)
}
\arguments{
\item{fetch}{a function that takes no arguments that, when called, attempts
to fetch the desired resource and returns the response}

\item{check_response}{a function that takes a fetch response as input and
either (a) stops if the response indicates an unsuccessful fetch or (b)
returns nothing; this prevents invalid fetch responses from being saved to
the cache.}

\item{cache_file_prefix}{length-1 \code{character} containing a filepath
prefix to use when creating files to cache the fetch response, or
\code{NULL} to disable caching}

\item{cache_invalidation_period}{length-1 \code{difftime}; the minimum amount
of time that must pass between the last cache update and a call to this
function for \code{fetch} to be called again rather than using the cache
result (unless \code{force_cache_invalidation=TRUE})}

\item{force_cache_invalidation}{\code{TRUE} to force a cache update, even if
the cache_invalidation_period has not passed; otherwise \code{FALSE}}

\item{silent}{\code{TRUE} or \code{FALSE} (default); \code{TRUE} to disable
messages about whether the cache is being used or not, or \code{FALSE} to
enable these messages}

\item{make_ancestor_directories}{Boolean; \code{TRUE} for any nonexistent
ancestor directories of cache files to be automatically created, or
\code{FALSE} for errors to be raised if these directories don't already
exist}
}
\description{
Fetch or read cache of a resource that potentially updates over time
}
